put model loading back into Keras class
